// constants/label.enum.ts
export enum LabelEnum {
  promoProduct_1 = "Творческий беспорядок",
  promoProduct_2 = "Игра престолов",
  promoProduct_3 = "Black&Red",
  noPromoProduct_1 = "Блокнот в точку",
  noPromoProduct_2 = "Кошечка Мари",
  noPromoProduct_3 = "Нотная тетрадь",
  noPromoProduct_4 = "Гусь. Дедлайн",
  noPromoProduct_5 = "Художник",
  noPromoProduct_6 = "Плюшевый кот",
  noPromoProduct_7 = "Дерево",
}